requirements.txt
Django==4.2.23
djangorestframework==3.16.1
django-crispy-forms==2.4
crispy-bootstrap5==2025.6
django-ckeditor==6.7.3
Pillow==11.3.0
stripe==12.5.0
python-decouple==3.8
requests==2.32.5

# Google OAuth dependencies
django-allauth==65.11.1
PyJWT==2.10.1
cryptography==45.0.6
